refactor ScrollTopButton
